service/requirements.txt
Django==3.1.3
djangorestframework==3.12.2
mysqlclient>=2.0.1
django-mysql==3.9
django-cors-headers==3.5.0
